pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ctxai"
version = "0.1.0"
description = "Ctx AI - Dynamic agentic AI framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web framework & async
    "a2wsgi==1.10.8",
    "flask[async]==3.0.3",
    "flask-basicauth==0.2.0",
    "python-socketio>=5.14.2",
    "uvicorn>=0.38.0",
    "wsproto>=1.2.0",
    "nest-asyncio==1.6.0",
    "fastapi>=0.110.0",

    # AI / LLM / ML
    "litellm>=1.79.3",
    "openai>=1.99.2",
    "langchain-core==0.3.49",
    "langchain-community==0.3.19",
    "langchain-unstructured==0.1.6",
    "sentence-transformers==3.0.1",
    "tiktoken==0.8.0",
    "faiss-cpu>=1.7.4,<1.8.1; sys_platform == 'darwin'",
    "faiss-cpu>=1.9.0; sys_platform != 'darwin'",

    # MCP / A2A
    "mcp==1.22.0",
    "fastmcp==2.13.1",
    "fasta2a==0.5.0",

    # Browser automation
    "browser-use==0.5.11",
    "playwright==1.52.0",

    # Speech
    "openai-whisper==20250625",
    "kokoro>=0.9.2",
    "soundfile==0.13.1",

    # Document processing
    "pypdf==6.0.0",
    "pymupdf==1.25.3",
    "pytesseract==0.3.13",
    "pdf2image==1.17.0",
    "unstructured[all-docs]==0.16.23",
    "unstructured-client==0.31.0",
    "langchain-unstructured==0.1.6",
    "markdown==3.7",
    "lxml_html_clean==0.3.1",
    "markdownify==1.1.0",
    "html2text>=2024.2.26",
    "beautifulsoup4>=4.12.3",
    "newspaper3k==0.2.8",

    # HTTP / Networking / Email
    "paramiko==3.5.0",
    "imapclient>=3.0.1",
    "boto3>=1.35.0",
    "exchangelib>=5.4.3",

    # Docker / Git / Projects
    "docker==7.1.0",
    "GitPython==3.1.43",
    "giturlparse==0.14.0",
    "flaredantic==0.1.5",
    "pathspec>=0.12.1",

    # Search / Scraping
    "duckduckgo-search==6.1.12",

    # Utilities
    "ansio==0.0.1",
    "inputimeout==1.0.4",
    "simpleeval==1.0.3",
    "python-dotenv==1.1.0",
    "pytz==2024.2",
    "webcolors==24.6.0",
    "crontab==1.0.1",
    "pydantic==2.11.7",
    "psutil>=7.0.0",

    # Platform-specific
    "pywinpty==3.0.2; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
    "pyinstrument==5.1.2",
]

[tool.uv]
# browser-use==0.5.11 pins openai==1.99.2, but litellm requires openai>=1.99.5
# and both work fine with newer openai at runtime. Use override to resolve.
# This replicates the behaviour of requirements2.txt (post-install override).
package = false
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
    "sys_platform == 'linux' and platform_machine == 'aarch64'",
    "sys_platform == 'win32' and platform_machine == 'AMD64'",
]
override-dependencies = [
    "openai>=1.99.5",
    "numba<0.61.0; sys_platform == 'darwin' and platform_machine == 'x86_64'",
]


[tool.hatch.build.targets.wheel]
packages = ["src/ctxai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py312"
